Moved materialView stuff into roundCtrl
